requirements.txt
autopep8==1.4.4
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
